+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- added new status for missing CPU ([#38](https://github.com/AlmostReliable/merequester/issues/38))
 - added the energy capability to the wireless terminal to allow charging with other
   mods ([Kolja](https://github.com/ko-lja)@[#51](https://github.com/AlmostReliable/merequester/pull/51))
 - fixed scroll bar drag control using wrong texture ([#49](https://github.com/AlmostReliable/merequester/issues/49))